requirements.txt
affine==2.4.0
attrs==23.1.0
azure-core==1.27.1
azure-functions==1.15.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
basemap==1.3.7
basemap-data==1.3.2
branca==0.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contextily==1.3.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.40.0
geographiclib==2.0
geopandas==0.13.2
geopy==2.3.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mercantile==1.2.1
msal==1.22.0
msal-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
patsy==0.5.3
Pillow==9.5.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
pyproj==3.5.0
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3
rasterio==1.3.7
requests==2.31.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
statsmodels==0.14.0
tenacity==8.2.2
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
xyzservices==2023.5.0
zipp==3.15.0
